Assignment>>Practice-Set-02

+++ Assignments/Practice_Set_02/Commission.java
@@ -8,10 +8,10 @@ public class Commission {
         Scanner in = new Scanner(System.in);
 
         System.out.print("Enter the amount:");
-        double amount = in.nextInt();
+        double amount = in.nextDouble();
 
         System.out.print("Enter your commission percentage:");
-        double commission_percent = in.nextInt();
+        double commission_percent = in.nextDouble();
 
         double commission = (amount/100)*commission_percent;
         System.out.println("You get a commission of Rs:"+commission);
